+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 - Added new `ValidateRejectCode` values to the `/v3/block_proposal` endpoint
 - Added `StateMachineUpdateContent::V1` to support a vector of `StacksTransaction` expected to be replayed in subsequent Stacks blocks
+- Updated `StackerDBListener` to now listens for signers' state machine updates and stores replay info to the side to enable a miner to perform transaction replay
 
 ### Changed